+++ ts/lib/url-parser.ts
@@ -8,7 +8,7 @@ const protocols: string[] = [
     'smb://'
 ];
 
-const multiExtensions: string[] = [
+const multiExtensionsPrefix: string[] = [
     'co',
     'com',
     'fr',
@@ -17,6 +17,41 @@ const multiExtensions: string[] = [
     'org',
     'edu',
     'net',
+    'gob',
+];
+
+const multiExtensions: string[] = [
+    // UK multi-extensions
+    'judiciary.uk',
+    'ltd.uk',
+    'me.uk',
+    'mod.uk',
+    'nhs.uk',
+    'nic.uk',
+    'parliament.uk',
+    'plc.uk',
+    'sch.uk',
+    'bl.uk',
+    'jet.uk',
+    'british-library.uk',
+    'nls.uk',
+
+    // Canadian provincial domains
+    // https://en.wikipedia.org/wiki/.ca#Third-level_.28provincial.29_and_fourth-level_.28municipal.29_domains
+    'ab.ca',
+    'bc.ca',
+    'mb.ca',
+    'nb.ca',
+    'nf.ca',
+    'nl.ca',
+    'ns.ca',
+    'nt.ca',
+    'nu.ca',
+    'on.ca',
+    'pe.ca',
+    'qc.ca',
+    'sk.ca',
+    'yk.ca',
 ];
 
 function domainIsIPv4(domainParts: string[]): boolean {
@@ -28,7 +63,8 @@ function domainIsIPv4(domainParts: string[]): boolean {
     } else {
         // ensure valid IP format (basic test)
         return (domainParts.length === 4) && domainParts.every( (domainPart: string) => {
-            return !isNaN(parseInt(domainPart, 10));
+            const byte = parseInt(domainPart, 10);
+            return !isNaN(byte) && byte <= 255;
         });
     }
 }
@@ -37,10 +73,22 @@ function domainIsIPv6(domainParts: string[]): boolean {
     // ex with url = 'http://[1080:0:0:0:8:800:200C:417A]'
     // domainParts = [ '[1080:0:0:0:8:800:200C:417A]' ]
     var ipv6 = domainParts[0];
-    return domainParts.length === 1 &&
+    const checkFormat = domainParts.length === 1 &&
            ipv6[0] === '[' &&
            ipv6[ ipv6.length - 1] === ']' &&
-           ipv6.split(':').length === 8;
+           ipv6.split(':').length >= 4 &&   // [2001:db8::1] is a valid IPv6 address and equal to [2001:db8:0:1:0:0:0:0]
+           ipv6.split(':').length <= 8;
+
+    const checkBytes = ipv6.slice(1, ipv6.length-1).split(':').every( (str: string) => {
+        if (!str) {
+            // In IPv6 protocol, this is a valid IPv6 address: [1080::::8:800:200C:417A]
+            str = '0';
+        };
+        const bytes = parseInt(str, 16);
+        return !isNaN(bytes) && bytes <= 0xFFFF;
+    });
+
+    return checkFormat && checkBytes;
 }
 
 function domainIsIP(domainParts: string[]): boolean {
@@ -110,7 +158,9 @@ export function extractRootDomain(url: string): string {
 
     if (domainParts.length <= 2) {
         return domain;
-    } else if (multiExtensions.indexOf(domainParts[domainParts.length - 2]) !== -1) {
+    } else if (multiExtensionsPrefix.indexOf(domainParts[domainParts.length - 2]) !== -1) {
+        domainParts = domainParts.slice(domainParts.length - 3);
+    } else if (multiExtensions.indexOf(domainParts.slice(domainParts.length - 2).join('.')) !== -1) {
         domainParts = domainParts.slice(domainParts.length - 3);
     } else {
         domainParts = domainParts.slice(domainParts.length - 2);
@@ -129,11 +179,22 @@ export function extractRootDomainName(url: string): string {
     }
 }
 
+export function extractSubDomainName(url: string): string {
+    const subDomainName = extractFullDomain(url)
+        .replace(extractRootDomain(url), '')
+        .replace(/\.$/, '');
+    if (!subDomainName || subDomainName === 'www') {
+        return null;
+    }
+    return subDomainName;
+}
+
 export interface ParsedUrl {
     url: string;
     fullDomain: string;
     rootDomain: string;
     rootDomainName: string;
+    subDomainName: string;
 }
 
 export function getParsedUrl(url: string): ParsedUrl {
@@ -141,7 +202,42 @@ export function getParsedUrl(url: string): ParsedUrl {
         url: url,
         fullDomain: extractFullDomain(url),
         rootDomain: extractRootDomain(url),
-        rootDomainName: extractRootDomainName(url)
+        rootDomainName: extractRootDomainName(url),
+        subDomainName: extractSubDomainName(url),
     };
 }
 
+export function isUrlWithIPv4(url: string): boolean {
+    const domain: string = extractFullDomain(url);
+    let domainParts: string[] = domain.split('.');
+    return domainIsIPv4(domainParts);
+}
+
+export function isUrlWithIPv6(url: string): boolean {
+    const domain: string = extractFullDomain(url);
+    let domainParts: string[] = domain.split('.');
+    return domainIsIPv6(domainParts);
+}
+
+export function isUrlWithIP(url: string): boolean {
+    const domain: string = extractFullDomain(url);
+    let domainParts: string[] = domain.split('.');
+    return domainIsIP(domainParts);
+}
+
+export function isUrlWithDomain(url: string): boolean {
+    // Regexps from the C++ file Desktop/CppLibrairies/SourceCode/KWMacUI/UI Applet/Tools/URLValidator.m
+    // Used in C++ in the function isValidUrlOrIP(url) to validate datacapture of a URL
+    const urlRegexps = [
+        /^(?:https?:\/\/)?(?:[a-z0-9\-_]{1,63}\.)+(?:[a-z0-9\-_]{1,63})(?::[0-9]{1,5})?(?:\/.*)?$/i,
+        /^(?:https?:\/\/)?(?:[a-z0-9\-_]{1,63})(?::[0-9]{1,5})?(?:\/.*)?$/i,
+        /^(?:https?:\/\/)?(?:[a-z0-9\-_]{1,63}\.)+(?:[a-z0-9\-_]{1,63})(?::[0-9]{1,5})?\s*$/i,
+        /^(?:https?:\/\/)?(?:[a-z0-9\-_]{1,63})(?::[0-9]{1,5})?\s*$/i,
+    ];
+    const urlRegex: RegExp = /^(?:https?:\/\/)?(?:[a-z0-9\\-_]{1,63}\\.)+(?:[a-z0-9\\-_]{1,63})(?::[0-9]{1,5})?(?:\/.*)?$/i;
+    return urlRegexps.some(urlRegexp => url.match(urlRegexp) !== null) && !isUrlWithIP(url);
+}
+
+export function isUrl(url: string): boolean {
+    return isUrlWithDomain(url) || isUrlWithIP(url);
+}